+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - `postgresqlreceiver`: add the receiver to available components (#7079)
 - `tanzuobservability exporter`: Support summary metrics (#7121)
 - Use Jaeger gRPC instead of Thrift in the docker-compose example (#7243)
+- `tanzuobservabilityexporter`: Support exponential histograms (#7127)
 
 ## 🛑 Breaking changes 🛑